Try enforcing np.uint64 dtype

+++ hstrat/frozen_instrumentation/_HereditaryStratigraphicAssemblageSpecimen.py
@@ -176,7 +176,7 @@ class HereditaryStratigraphicAssemblageSpecimen:
             A numpy array containing ranks of differentia entries, including
             null entries for differentia that are not retained.
         """
-        return self._data.index.array.to_numpy(copy=False)
+        return self._data.index.array.to_numpy(copy=False, dtype=np.uint64)
 
     def GetRankAtColumnIndex(
         self: "HereditaryStratigraphicAssemblageSpecimen",
